Crate key_server_cli

Implementation of our LockKeeper server.

Logging

Our server supports logging using the tracing crate.

Main tracing Concepts.

We utilize the following concept from tracing:

Events: These are the individual logging events that are called as our program executes. Events have an info level attached to them. You may create a new event by using the appropriate macro event macro from tracing, e.g:

info!("This is a logging event!");

You may add runtime values to our events as well:

info!("Logging config settings: {:?}", config.logging);

This uses the same syntax as Rust string formatting.

Spans represent logical regions of code, usually a function. All _event_s within a span will have additional span data attached to the event. This allows us to add context-relevant data to an event without having to explicitly add this data to our event. For example an event like info!(This operation completed successfully!) within our span will look like:

  2022-12-06T18:57:40.022422Z  INFO lock_keeper_key_server::server::operation: This operation completed successfully!
    at lock-keeper-key-server/src/server/operation.rs:72
    in lock_keeper_key_server::server::operation::handle_request with request_id: "52ad8008-1eb9-4ad5-8bce-0344a958fd1e", action: "Authenticate"

This event has the handle_request span attached to it. Additionally, spans may contain fields. These fields attached relevant data to all events within our span. In the log message above, we see our span has fields request_id and action attached.

Adding your Own Span.

The best way to create your own span is using the instrument macro. This macro will automatically create a new span when the instrumented function is called. This span will have the name of the function and the module path as the target (this can be used for filtering by certain targets).

Please follow the following template for instrumenting your function with spans:

#[instrument(skip(sensitive_arg), err(Debug))]
fn foo(arg1: _, arg2: _, sensitive_arg SuperSecret) -> _ {}

By default, instrument will create a new span where every argument to a function is a span field. This is not desirable for our key server as we want to avoid logging sensitive data! Make sure to skip sensitive arguments with the skip(sensitive_arg_1, sensitive_arg_2) option. You can also use the skip_all option if none of the function arguments need to be logged.

Finally, the err(Debug) option tells tracing that any Result::Err(_) returned from this function should be logged as events (with the default level of INFO). This is useful for logging where errors originated from in our code. The Debug part tells tracing to use the Debug implementation of this type for formatting, instead of the default Display.

Note that by default, return values of functions are not recorded. You can add the ret option to instrument if you wish to record the Ok(_) return value as a field.

tracing relies on the Debug implementation of types for printing values. Your type must implement Debug if you wish to include it as a field.

Recording Fields Not Immediately Available as Arguments.

Sometimes we want to add a field to a span where the field data isn’t immediately available as an argument to the function. Consider the following example:

#[instrument(skip_all, err(Debug), fields(action, user_id, request_id))]
    async fn handle_request(
        self,
        mut context: Context<DB>,
        request: Request<Streaming<Message>>,
    ) -> Result<_, _> {

        let request_id = Uuid::new_v4();
        logging::record_field("request_id", &request_id);
        info!("Handling new client request.");

        ... // Lots more things happen afterwards.
    }

Here we specify the fields action, user_id, request_id even though none of these are function arguments! When the function body executes, we create a request_id which we record using our [record_field][lock_keeper::infrastructure::logging::record_field] function. So our event info!("Handling new client request."); will have the request_id field attached:

  2022-12-06T18:57:39.667194Z  INFO
lock_keeper_key_server::server::operation: Handling new client request.
    at lock-keeper-key-server/src/server/operation.rs:48
    in lock_keeper_key_server::server::operation::handle_request with request_id: "52ad8008-1eb9-4ad5-8bce-0344a958fd1e"

Notice we have not recorded the action and request_id fields by the time this event is logged, so these fields are not included. Span fields can be recorded later as long as they are declared along with the span.

Note: While it is possible to create your own spans, it is not recommended. The instrument macro handles spans for asynchronous functions. Ensuring spans are properly kept track of, even when .awaiting across spans.

Which Arguments Should be Recorded as Fields?

In general, we should only include fields for our span that provide important information that should be attached to every event that happens within that span.

Fields are useful for attaching data to every event without having to explicitly pass that data to the event.

Note: We should be careful to avoid recording any sensitive data as fields!

Which Function Should be instrumented?

We want to instrument functions which either:

Logically represent an important step in our server processing.

Record useful fields to be attached to event within that span.

Have return values which are useful to log.

Instrumenting multiple functions which call each other will create a useful “backtrace-like” context which allows us to understand the calling context of an event. For example:

  2022-12-06T18:57:39.669808Z  INFO
lock_keeper_key_server::operations::authenticate: Starting authentication protocol.
    at lock-keeper-key-server/src/operations/authenticate.rs:38
    in lock_keeper_key_server::operations::authenticate::operation
    in lock_keeper_key_server::server::operation::handle_request with request_id: "52ad8008-1eb9-4ad5-8bce-0344a958fd1e", action: "Authenticate"

We want to avoid instrumenting uninteresting functions, e.g. functions which represent an implementation detail instead of a logical step in our server execution.

Logging Levels

So far our logging documentation has focused on the default tracing level of INFO. Events may be added for the following tracing levels: error, warn, info, debug, and trace. Here we give a brief overview of what falls under every level:

error: Report unexpected errors or failures in the system.

warn: Warn of unexpected errors, unforeseen program states, or possible issues that are not necessarily fatal, but should be logged so they may be debugged later.

info: For high-level, relevant, events to the execution of the code. This is our default logging level and should produce useful messages for developers.

debug: More verbose, good to have information for debugging purposes.

trace: Very verbose information that allows you to trace the execution of a program at a fine granularity, e.g. program control flow.

Modules

config 🔒

Config for key server binary.

Structs

Cli

LoggingGuards 🔒

Object representing our logging. Should be kept around as our logging writers return guards that should live for the lifetime of the program. Do not do anything with the guards. Just make sure they are not dropped!

Functions

get_database_config 🔒

get_paths 🔒

Return the path directory and the file name. Needed for passing to tracing_appender.

get_session_cache_config 🔒

init_logging 🔒

Initialize our logging with different logging layers:

main

our_targets_filter 🔒

Create filters for logging events originating from our lock_keeper* crates.

run_main