Function key_server_cli::init_logging

pub(crate) fn init_logging(
    all_logs: &Path,
    server_logs: &Path
) -> Result<Logging, LockKeeperServerError>

Initialize our logging with three different logging layers:

Log all INFO-level messages (or higher) from our key_server* crates to standard out. 2) Log all messages (TRACE or higher) from our key_server* crates to the path specified by server_logs.

Log all messages ((TRACE or higher) from any crate to the path specified by all_logs.

Returns an object which should be kept around for the lifetime of the program.